requirements.txt
django==4.2.1
python-decouple==3.8
django-debug-toolbar==4.0.0
django-cors-headers==3.14.0
django-phonenumber-field==7.1.0
pillow==9.5.0
djangorestframework==3.14.0
drf_spectacular==0.26.2
djangorestframework-simplejwt==5.2.2
django-filter==23.2
celery==5.2.7
channels[daphne]==4.0.0
django-import-export==3.2.0
psycopg2-binary==2.9.6
channels-redis==4.1.0
